+++ CHANGELOG.md
@@ -6,3 +6,5 @@ All notable changes to the "zambra-s-snippets" extension will be documented in t
 
 - 0.0.15 -> Add snippet 'Send Mail'
 - 0.0.16 -> Add snippet 'implementBatch'
+- 0.0.17 -> Fix image and git reporitory
+- 0.0.18 -> Add commands 'Force Apex Deploy' and 'Force Lightning Web Component Deploy'